+++ core/src/main/java/io/github/package_game_survival/entidades/seres/enemigos/InvasorArquero.java
@@ -1,6 +1,8 @@
 package io.github.package_game_survival.entidades.seres.enemigos;
 
 import com.badlogic.gdx.graphics.g2d.TextureAtlas;
+import io.github.package_game_survival.entidades.objetos.Carne;
+import io.github.package_game_survival.entidades.objetos.CarnePodrida;
 import io.github.package_game_survival.entidades.seres.jugadores.Jugador;
 import io.github.package_game_survival.habilidades.AtaqueAranazo;
 import io.github.package_game_survival.managers.Assets;
@@ -16,11 +18,9 @@ public class InvasorArquero extends Enemigo {
         // AJUSTE DE IA: Rango medio-largo
         this.rangoAtaque = 150f;
 
-        // AJUSTE DE ARMA:
-        // Cooldown: 3.5s
-        // Casteo: 1.2s (Lento, tarda en tensar el arco)
-        // Daño: 18 (Pega duro)
-        // Rango: 150f
         this.habilidadPrincipal = new AtaqueAranazo(3.5f, 1.2f, 18, 150f, 20f, Jugador.class);
+
+        this.agregarDrop(CarnePodrida.class, 1f);
+        this.agregarDrop(Carne.class, 0.6f);
     }
 }